(refactor): Removed mapstructure as it is no longer maintained

+++ parser/index_test.go
@@ -28,14 +28,14 @@ func TestTakeBinaryData(t *testing.T) {
 		Type: EVENT,
 		Data: []any{
 			"string data",
-			map[string]any{"_placeholder": true, "num": 0},
+			map[string]any{"_placeholder": true, "num": float64(0)},
 			map[string]any{
 				"key":  "value",
-				"file": map[string]any{"_placeholder": true, "num": 1},
+				"file": map[string]any{"_placeholder": true, "num": float64(1)},
 			},
-			map[string]any{"_placeholder": true, "num": 2},
+			map[string]any{"_placeholder": true, "num": float64(2)},
 			types.NewStringBuffer([]byte{0x07, 0x08, 0x09}),
-			map[string]any{"_placeholder": true, "num": 3},
+			map[string]any{"_placeholder": true, "num": float64(3)},
 		},
 		Attachments: new(uint64),
 	}
@@ -81,12 +81,12 @@ func TestTakeBinaryDataErrorHandling(t *testing.T) {
 		Type: EVENT,
 		Data: []any{
 			"string data",
-			map[string]any{"_placeholder": true, "num": 0},
+			map[string]any{"_placeholder": true, "num": float64(0)},
 			map[string]any{
 				"key":  "value",
-				"file": map[string]any{"_placeholder": true, "num": 1},
+				"file": map[string]any{"_placeholder": true, "num": float64(1)},
 			},
-			map[string]any{"_placeholder": true, "num": 2},
+			map[string]any{"_placeholder": true, "num": float64(2)},
 			types.NewStringBuffer([]byte{0x07, 0x08, 0x09}),
 		},
 		Attachments: new(uint64),
@@ -115,12 +115,12 @@ func TestCleanUp(t *testing.T) {
 		Type: EVENT,
 		Data: []any{
 			"string data",
-			map[string]any{"_placeholder": true, "num": 0},
+			map[string]any{"_placeholder": true, "num": float64(0)},
 			map[string]any{
 				"key":  "value",
-				"file": map[string]any{"_placeholder": true, "num": 1},
+				"file": map[string]any{"_placeholder": true, "num": float64(1)},
 			},
-			map[string]any{"_placeholder": true, "num": 2},
+			map[string]any{"_placeholder": true, "num": float64(2)},
 			types.NewStringBuffer([]byte{0x07, 0x08, 0x09}),
 		},
 		Attachments: new(uint64),
